requirements.txt
events-jimobama @ git+https://github.com/miljimo/events.git@02e5be98bf50e65fc6c07089238aa16615853fa4
numpy==1.18.4
Pillow==7.1.2
six==1.15.0
wxPython==4.1.0
